Add example for multicomponent cabal projects (#36)
I struggled through this configuration recently and thought it might be
useful to include in the example configurations.

If there is any interest in an update to the documentation in general
I'd like to add some Diataxis organised documentation on using this
plugin with cabal.  Let me know if that's a reasonable extension and
I'll submit more pull requests.  Otherwise, this simple addition will
suffice.